requirements.txt
# SparkJAR Crew API Requirements
# Based on CrewAI 0.148.0 and actual codebase usage
# ChromaDB is included as a dependency of CrewAI (client mode)

# Core CrewAI - MUST BE LATEST
crewai==0.148.0
crewai-tools>=0.51.1

# API Framework - Compatible versions
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart>=0.0.6

# Database - Compatible with CrewAI
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg>=0.30.0
pgvector==0.4.1
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.9.0

# Vector Databases - Compatible with CrewAI 0.148.0
chromadb==1.0.15

# LLM and Embeddings - Compatible versions
openai>=1.13.3
tiktoken>=0.5.2

# Utilities - Compatible with CrewAI
python-dotenv>=1.0.0
pydantic==2.11.7
pyyaml>=6.0.0
httpx>=0.25.0
requests>=2.31.0
cachetools>=5.3.0

# Google Drive Integration - Compatible versions
google-api-python-client==2.176.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
fpdf2>=2.7.0
Pillow>=10.0.0

# Email
sendgrid>=6.11.0

# Chat Interface Dependencies
redis>=5.0.0
structlog>=24.1.0
prometheus-client>=0.19.0

# Schema Validation
jsonschema>=4.20.0

# Browser automation (optional)
playwright>=1.49.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0